+++ src/Transport/TransportFactory.php
@@ -36,8 +36,7 @@ final class TransportFactory implements TransportFactoryInterface
 
     public const RESETTABLE = 'resettable';
 
-    /** @var ContainerInterface */
-    private $container;
+    private ContainerInterface $container;
 
     public function __construct(ContainerInterface $container)
     {
@@ -89,11 +88,9 @@ final class TransportFactory implements TransportFactoryInterface
     private function addDocumentEnhancers(Connection $connection, array $options): void
     {
         foreach ($options[self::DOCUMENT_ENHANCERS] as $name) {
-            if ($this->isServiceDefinition($name)) {
-                $enhancer = $this->container->get(ltrim($name, '@'));
-            } else {
-                $enhancer = new $name();
-            }
+            $enhancer = $this->isServiceDefinition($name)
+                ? $this->container->get(ltrim($name, '@'))
+                : new $name();
 
             if (! $enhancer instanceof DocumentEnhancer) {
                 throw new \InvalidArgumentException('Expecting DocumentEnhancer, got ' . get_debug_type($enhancer));
@@ -152,13 +149,23 @@ final class TransportFactory implements TransportFactoryInterface
     }
 
     /**
-     * @param string[] $enhancers
+     * @param mixed $enhancers
      *
      * @throws \InvalidArgumentException If any of the document_enhancers values is not valid
+     *
+     * @phpstan-assert array<array-key, string> $enhancers
      */
-    private function validateDocumentEnhancers(array $enhancers): void
+    private function validateDocumentEnhancers($enhancers): void
     {
+        if (! is_array($enhancers)) {
+            throw new \InvalidArgumentException('Document enhancers should array, got ' . get_debug_type($enhancers));
+        }
+
         foreach ($enhancers as $name) {
+            if (! is_string($name)) {
+                throw new \InvalidArgumentException('Document enhancers should be strings, got ' . get_debug_type($name));
+            }
+
             if ($this->isServiceDefinition($name)) {
                 continue;
             }