+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+* Improve config & serialization validation thanks to PHPStan v2 upgrade
+
 ## 1.6.0 (2025-04-23)
 * Support `ext-mongodb` and `mongodb/mongodb` v2